Logics for routers separate files
