+++ routers/users.js
@@ -1,27 +1,12 @@
 const router = require('express').Router();
-const fsPromises = require('fs').promises;
-const findUser = require('./finduser.js');
-
-const users = fsPromises.readFile('./data/users.json', { encoding: 'utf8' });
+const { getUsersList, getUserById } = require('./users-logics');
 
 router.get('/', (req, res) => {
-  users
-    .then(data => {
-      res.send(data);
-    })
-    .catch(err => {
-      console.log(`Проблема с получением данных о пользователях, ошибка: ${err}`);
-    });
+  getUsersList(res);
 });
 
 router.get('/:id', (req, res) => {
-  users
-    .then(data => {
-      findUser(JSON.parse(data), res, req);
-    })
-    .catch(err => {
-      console.log(`Проблема с получением данных о пользователе, ошибка: ${err}`);
-    });
+  getUserById(req, res);
 });
 
 module.exports = router;
